% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_company}
\alias{diario_get_company}
\title{Get company details}
\usage{
diario_get_company()
}
\value{
A list containing company details.
}
\description{
This function retrieves company details from the Diario API.
}
\examples{
\dontrun{
company <- diario_get_company()
}
}
