% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_entities}
\alias{diario_get_entities}
\title{Get all registered entities (cadastros)}
\usage{
diario_get_entities()
}
\value{
A list containing the entities data.
}
\description{
This function retrieves all registered entities from the Diario API.
}
\examples{
\dontrun{
entities <- diario_get_entities()
}
}
