% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_project_details}
\alias{diario_get_project_details}
\title{Get details of a specific project}
\usage{
diario_get_project_details(project_id)
}
\arguments{
\item{project_id}{A valid non-empty string with the project ID.}
}
\value{
A list containing the project details.
}
\description{
This function retrieves details of a specific project by its ID from the Diario API.
}
\examples{
\dontrun{
project <- diario_get_project_details("66face5fe26175e0a904d398")
}
}
