% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_projects}
\alias{diario_get_projects}
\title{Get list of projects (obras)}
\usage{
diario_get_projects()
}
\value{
A tibble containing the projects data.
}
\description{
This function retrieves a list of projects from the Diario API.
}
\examples{
\dontrun{
projects <- diario_get_projects(query = list(status = "active"))
}
}
