% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_report_details}
\alias{diario_get_report_details}
\title{Get details of a specific report}
\usage{
diario_get_report_details(project_id, report_id)
}
\arguments{
\item{project_id}{A valid non-empty string with the project ID.}

\item{report_id}{A valid non-empty string with the report ID.}
}
\value{
A list containing the report details.
}
\description{
This function retrieves details of a specific report by report ID within a project.
}
\examples{
\dontrun{
report <- diario_get_report_details("6717f864d163f517ae06e242", "67648080f0971de9d00324c2")
}
}
