% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_retrieve_token}
\alias{diario_retrieve_token}
\title{Retrieve the API token for Diario}
\usage{
diario_retrieve_token()
}
\value{
A character string containing the API token.
}
\description{
This function retrieves the stored authentication token using the \code{keyring} package.
}
\examples{
\donttest{
\dontrun{
token <- diario_retrieve_token()
}
}
}
