% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{diario_get_task_list}
\alias{diario_get_task_list}
\title{Get the task list of a specific project}
\usage{
diario_get_task_list(project_id)
}
\arguments{
\item{project_id}{A valid non-empty string with the project ID.}
}
\value{
A tibble containing the task list.
}
\description{
This function retrieves the task list of a specific project by its ID.
}
\examples{
\dontrun{
tasks <- diario_get_task_list("66cf438223aa80386306e647")
}
}
